+++ src/components/Task/index.jsx
@@ -1,5 +1,7 @@
+import { useState } from 'react';
 import { AiOutlineDelete, AiOutlineEdit } from 'react-icons/ai'
 import { dateFormat } from '../../util/format';
+import EditTaskModal from '../EditTaskModal';
 
 import { 
   Table,
@@ -12,28 +14,39 @@ import {
 export function Task(prop) {
   const { id, task, status, createAt } = prop.task;
   const destroyTask = prop.destroyTask
-  const modalOpen = prop.EditTaskModalOpen
+  const [isEditTaskModalOpen, setisEditTaskModalOpen] = useState(false);
   const dateFormated = dateFormat(createAt);
+
 
   function handleDeleteTask() {
     destroyTask(id)
   }
 
+  function handleOpenEditTaskModalClose() {
+    setisEditTaskModalOpen(false);
+  }
+
   return (
-    <Table>
-      <Description>{task}</Description>
-      <Status status={status}>{status}</Status>
-      <div>{dateFormated}</div>
-      <Edit>
-        <button onClick={modalOpen}>
-          <AiOutlineEdit />
-        </button>
-      </Edit>
-      <Delete>
-        <button onClick={handleDeleteTask}>
-          <AiOutlineDelete />
-        </button>
-      </Delete>
-    </Table>
+    <>
+      <Table>
+        <Description>{task}</Description>
+        <Status status={status}>{status}</Status>
+        <div>{dateFormated}</div>
+        <Edit>
+          <button onClick={() => setisEditTaskModalOpen(true)}>
+            <AiOutlineEdit />
+          </button>
+        </Edit>
+        <Delete>
+          <button onClick={handleDeleteTask}>
+            <AiOutlineDelete />
+          </button>
+        </Delete>
+      </Table>
+      <EditTaskModal
+        isOpen={isEditTaskModalOpen}
+        onRequestClose={handleOpenEditTaskModalClose}           
+      />
+    </>
   );
 };
